% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et-book.R
\docType{data}
\name{font_et}
\alias{font_et}
\title{ET Book fonts R variable aliases}
\format{length 5 character vector. Values are the
font names, and names are the description of the font.}
\usage{
font_et
}
\description{
ET Book fonts
}
\details{
ET Book (ETBembo font-family) comprises five fonts:
\itemize{
\item \code{"ETBembo RomanLF"}: Roman (lining figures)
\item \code{"ETBembo RomanOSF"}: Roman (oldstyle figures)
\item \code{"ETBembo SemiBoldOSF"}: Semi-bold (oldstyle figures)
\item \code{"ETBembo BoldLF"}: Bold (lining figures)
\item \code{"ETBembo DisplayItalic"}: Display Italic (oldstyle figures)
}

For headings, the sans serif font, Gill Sans is used.
Since there is no Gill Sans variant with an open license, the similar
Lato font is included with this package.
}
\examples{
font_et
}
\keyword{datasets}
