% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-quartileplot.R
\docType{data}
\name{geom_quartileplot}
\alias{geom_quartileplot}
\alias{GeomQuartilePlot}
\title{A box and whisker's plot (in the style of Tufte)}
\usage{
geom_quartileplot(mapping = NULL, data = NULL, stat = "fivenumber",
  position = "dodge", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, middle.line = FALSE, whisker.line = TRUE,
  hoffset = 0.01, fatten = 1, gapsize = 0.01, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
Tufte \emph{The Visual Display of Quantitative Information} (Ch 6)
proposes several revisions of the box plot.
These variants compactly display the distribution of a continuous variable.
It visualises five summary statistics: the median, two hinges (25\% and 75\% quartiles)
and two whiskers (minimum, maximum).
}
\examples{
library("ggplot2")
library("extrafont")
extrafont::loadfonts(quiet = TRUE)
extrafont::loadfonts("postscript", quiet = TRUE)

p <- ggplot(mtcars, aes(factor(cyl), mpg))
# with a point for the median and lines for whiskers
p + geom_quartileplot()
# with a line for the interquartile range and points for whiskers
p + geom_quartileplot(middle.line = TRUE, whisker.line = FALSE)
# with a wide line for the interquartile range and lines for whiskers
p + geom_quartileplot(middle.line = TRUE, hoffset = 0, fatten = 2)
# with an offset line for the interquartile range and lines for whiskers
p + geom_quartileplot(middle.line = TRUE, hoffset = 0.01, fatten = 1)
# combined with theme_tufte
p + geom_quartileplot() +
  theme_tufte() +
  theme(axis.ticks.x = element_blank())
}
\references{
Tufte, Edward R. (2001) The Visual Display of
  Quantitative Information, Chapter 6.
}
\seealso{
\code{\link{geom_boxplot}}

\code{\link{geom_boxplot}} and \code{\link{geom_pointrange}}.

Other geom tufte: \code{\link{geom_rangeframe}}
}
\keyword{datasets}
