% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_breaks.R
\name{range_breaks}
\alias{range_breaks}
\title{Pretty axis breaks inclusive of extreme values}
\usage{
range_breaks(x, n = 5, Q = c(1, 5, 2, 2.5, 4, 3), w = c(0.25, 0.2, 0.5,
  0.05))
}
\arguments{
\item{x}{Vector of numeric values}

\item{n}{desired number of breaks}

\item{Q}{set of nice numbers}

\item{w}{weights applied to the four optimization components (simplicity, coverage, density, and legibility)}
}
\value{
For \code{range_breaks}, a function which takes a single argument,
  a vector of data, which returns the vector of axis label locations.
}
\description{
This function returns pretty axis breaks that always include the extreme  values of the data.
This works by calling the extended Wilkinson algorithm (Talbot et. al, 2010), constrained to solutions interior to the data range.
Then, the minimum and maximum labels are moved to the minimum and maximum of the data range.
}
\references{
Talbot, J., Lin, S., Hanrahan, P. (2010) An Extension of Wilkinson's Algorithm for Positioning Tick Labels on Axes, InfoVis 2010.
}
\author{
Justin Talbot, Jeffrey B. Arnold, Baptiste Auguie
}
