% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tufte.R
\docType{data}
\name{tufte_theme_fonts}
\alias{tufte_theme_fonts}
\title{Font options to use with the Tufte Theme}
\format{A character vector.}
\usage{
tufte_theme_fonts
}
\description{
List of font or font-families to use with \code{theme_tufte}.
It starts with the preferred ETBembo font used in Tufte's book along with
fallback options.
}
\details{
This list is derived from the list of fonts used in
\href{tufte-css}{https://github.com/edwardtufte/tufte-css}
}
\examples{
tufte_theme_fonts
}
\keyword{datasets}
